requirements_test.txt
black==24.8.0
codespell==2.4.1
flake8==3.9.2
homeassistant==2025.1.0
mypy==0.901
pydocstyle==6.1.1
pylint-strict-informational==0.1
pylint==2.8.2
pytest-asyncio
pytest-cov
pytest_homeassistant_custom_component
ipdb